requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0 